requirements.txt
altair
numpy
pandas
pydeck
streamlit
plotly.express